requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Training & Evaluation
scikit-learn>=1.3.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Web Interface
gradio>=4.0.0

# Utilities
PyYAML>=6.0
timm>=0.9.0  # For EfficientNet pretrained models
albumentations>=1.3.0  # Advanced data augmentation
gdown>=4.7.0  # Google Drive downloads

# Optional (for deployment)
flask>=2.3.0
gunicorn>=21.0.0
